NGWPC-6905 - Fix PR comments/rename nwm_module tags
