+++ app/routers/nwm_modules/router.py
@@ -29,7 +29,7 @@ topmodel_router = APIRouter(prefix="/modules/topmodel")
 topoflow_router = APIRouter(prefix="/modules/topoflow")
 
 
-@sft_router.get("/", tags=["HF Modules"])
+@sft_router.get("/", tags=["NWM Modules"])
 async def get_sft_ipes(
     identifier: str = Query(
         ...,
@@ -71,7 +71,7 @@ async def get_sft_ipes(
     )
 
 
-@snow17_router.get("/", tags=["HF Modules"])
+@snow17_router.get("/", tags=["NWM Modules"])
 async def get_snow17_ipes(
     identifier: str = Query(
         ...,
@@ -113,7 +113,7 @@ async def get_snow17_ipes(
     )
 
 
-@smp_router.get("/", tags=["HF Modules"])
+@smp_router.get("/", tags=["NWM Modules"])
 async def get_smp_ipes(
     identifier: str = Query(
         ...,
@@ -155,7 +155,7 @@ async def get_smp_ipes(
     )
 
 
-@lstm_router.get("/", tags=["HF Modules"])
+@lstm_router.get("/", tags=["NWM Modules"])
 async def get_lstm_ipes(
     identifier: str = Query(
         ...,
@@ -190,7 +190,7 @@ async def get_lstm_ipes(
     )
 
 
-@lasam_router.get("/", tags=["HF Modules"])
+@lasam_router.get("/", tags=["NWM Modules"])
 async def get_lasam_ipes(
     identifier: str = Query(
         ...,
@@ -241,7 +241,7 @@ async def get_lasam_ipes(
     )
 
 
-@noahowp_router.get("/", tags=["HF Modules"])
+@noahowp_router.get("/", tags=["NWM Modules"])
 async def get_noahowp_ipes(
     identifier: str = Query(
         ...,
@@ -276,7 +276,7 @@ async def get_noahowp_ipes(
     )
 
 
-@sacsma_router.get("/", tags=["HF Modules"])
+@sacsma_router.get("/", tags=["NWM Modules"])
 async def get_sacsma_ipes(
     identifier: str = Query(
         ...,
@@ -318,7 +318,7 @@ async def get_sacsma_ipes(
     )
 
 
-@troute_router.get("/", tags=["HF Modules"])
+@troute_router.get("/", tags=["NWM Modules"])
 async def get_troute_ipes(
     identifier: str = Query(
         ...,
@@ -353,7 +353,7 @@ async def get_troute_ipes(
     )
 
 
-@topmodel_router.get("/", tags=["HF Modules"])
+@topmodel_router.get("/", tags=["NWM Modules"])
 async def get_topmodel_ipes(
     identifier: str = Query(
         ...,
@@ -389,7 +389,7 @@ async def get_topmodel_ipes(
 
 
 # TODO - Restore endpoint once the generation of IPEs for TopoFlow is possible/implemented
-# @topoflow_router.get("/", tags=["HF Modules"])
+# @topoflow_router.get("/", tags=["NWM Modules"])
 # async def get_topoflow_ipes(
 #     identifier: str = Query(
 #         ...,
@@ -424,7 +424,7 @@ async def get_topmodel_ipes(
 #     )
 
 
-@topoflow_router.get("/albedo", tags=["HF Modules"])
+@topoflow_router.get("/albedo", tags=["NWM Modules"])
 async def get_albedo(
     landcover_state: Albedo = Query(
         ...,
